microservices/fetch-repo-from-blob/requirements.txt
azure-storage-blob
fastapi
azure.core
uvicorn